# rest-api/src/main/resources/swagger/animal-super-store.yaml
---
swagger: "2.0"
info:
  description: "This is an Animal Super Store REST API."
  version: "1.0.0"
  title: "Animal Super Store"
host: "localhost:8080"
basePath: "/api"
tags:
  - name: "animal-api"
    description: "Application about Animals"
schemes:
  - "http"
paths:

  /animal:
    post:
      tags:
        - "animal-rest-endpoint"
      summary: "Add a new animal to the Super Store"
      description: "Creates a new Animal in the Super Store"
      operationId: "addAnimal"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Animal object that needs to be added to the Super Store"
          required: true
          schema:
            $ref: "#/definitions/AddAnimalDto"
      responses:
        '201':
          description: "Successfully created"
        '404':
          description: "Animal not found in third-party source"
        '405':
          description: "Invalid input"
    get:
      tags:
        - "animal-rest-endpoint"
      summary: "Get all animals available in the Animal Super Store"
      description: "Returns all animals available in the Animal Super Store"
      operationId: "getAllAnimals"
      produces:
        - "application/json"
      responses:
        '200':
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AnimalDto"

  /thirdParty/animal/findByStatus:
    get:
      tags:
        - "third-party-rest-endpoint"
      summary: "Finds Animals by status from all third party providers"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findThirdPartyAnimalByStatus"
      produces:
        - "application/json"
      parameters:
        - name: "status"
          in: "query"
          description: "Status values that need to be considered for filter"
          required: true
          type: "string"
          enum:
            - "available"
            - "pending"
            - "sold"
          default: "available"
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AnimalDto"
        400:
          description: "Invalid status value"

  /thirdParty/animal/{animalId}:
    get:
      tags:
        - "third-party-rest-endpoint"
      summary: "Find animal by ID in all third party providers"
      description: "Returns a single animal"
      operationId: "getThirdPartyAnimalById"
      produces:
        - "application/json"
      parameters:
        - name: "animalId"
          in: "path"
          description: "ID of animal to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        '200':
          description: "successful operation"
          schema:
            $ref: "#/definitions/AnimalDto"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Animal not found"

definitions:
  AnimalDto:
    type: "object"
    required:
      - "name"
      - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        items:
          type: "string"
      source:
        type: "string"
        description: "third-party source"
        enum:
          - "swagger_petstore"
      externalStatus:
        type: "string"
        description: "animal status in the store"
        enum:
          - "available"
          - "pending"
          - "sold"
          - "not_applicable"
      internalStatus:
        type: "string"
        description: "animal status in the store"
        enum:
          - "available"
          - "pending"
          - "sold"
          - "importing"

  AddAnimalDto:
    type: "object"
    required:
      - "id"
      - "source"
    properties:
      id:
        type: "integer"
        format: "int64"
      source:
        type: "string"
        description: "third-party source"
        enum:
          - "swagger_petstore"